refactor(core,service): migrate image service to multimodal with unified observation types
- Migrate image service to new multimodal service architecture
- Replace generic DataContentBlock with specific message content types
  (MessageContentImageUrl, MessageContentText) for better type safety
- Unify AgentObservation type to support both string and content block arrays
- Update OpenAI agent to work with unified observation types
- Add multimodal service plugin to dynamic import processor

BREAKING CHANGE: AgentObservation type definition has changed from
`string | DataContentBlock[]` to `(MessageContentImageUrl | MessageContentText)[] | string`
