--- packages/service-image/src/locales/en-US.schema.yml
@@ -1,14 +1,0 @@
-$inner:
-    - $desc: Model Options
-      model: 'Select a model that supports image input (e.g.: gpt-4o, gemini-2.0-flash)'
-      enableImageTool: Enable registering the image reading tool
-      imagePrompt: Image input system prompt
-      imageInsertPrompt: Image description insertion prompt
-      gifStrategy:
-        $desc: GIF frame extraction strategy
-        $inner:
-          - Extract only first frame
-          - Extract first N frames
-          - Extract average N frames
-
-      gifFrameCount: Number of frames to extract (only effective for "head N frames" or "average N frames" strategy)
