--- packages/service-image/src/locales/zh-CN.schema.yml
@@ -1,16 +1,0 @@
-$inner:
-    - $desc: 模型选项
-      model: '选择支持图片输入的模型（如：gpt-4o, gemini-2.0-flash)'
-      enableImageTool: 允许注册为读图工具
-      imagePrompt: 图片输入的系统提示词。
-      imageInsertPrompt: 图片描述插入到具体消息里的提示词。
-      gifStrategy:
-        $desc: GIF 图片帧提取策略
-        $inner:
-          - 只获取第一帧
-          - 获取前 N 帧
-          - 平均获取 N 帧
-
-      gifFrameCount: 提取的帧数（仅在选择"前 N 帧"或"平均 N 帧"策略时有效）
-
-
