+++ packages/service-multimodal/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "koishi-plugin-chatluna-image-service",
-  "description": "image service for chatluna",
-  "version": "1.3.4",
+  "name": "koishi-plugin-chatluna-multimodal-service",
+  "description": "multimodal service for chatluna",
+  "version": "1.3.0",
   "main": "lib/index.cjs",
   "module": "lib/index.mjs",
   "typings": "lib/index.d.ts",
