+++ packages/service-multimodal/src/locales/en-US.schema.yml
@@ -1,0 +1,16 @@
+$inner:
+    - $desc: Basic Settings
+      enableMultimodaTool: Enable multimodal input tool
+
+    - $desc: Image Multimodal Config
+      imageModel: 'Select an image model that supports image input (e.g.: gpt-4o, gemini-2.0-flash)'
+      imagePrompt: Image input system prompt
+      imageInsertPrompt: Image description insertion prompt
+      gifStrategy:
+          $desc: GIF frame extraction strategy
+          $inner:
+              - Extract only first frame
+              - Extract first N frames
+              - Extract average N frames
+
+      gifFrameCount: Number of frames to extract (only effective for "head N frames" or "average N frames" strategy)
