+++ packages/service-multimodal/src/locales/zh-CN.schema.yml
@@ -1,0 +1,16 @@
+$inner:
+    - $desc: 基础设置
+      enableMultimodaTool: 启用多模态输入工具
+
+    - $desc: 图片多模态配置
+      imageModel: '选择支持图片输入的图片模型（如：gpt-4o, gemini-2.0-flash)'
+      imagePrompt: 图片输入的系统提示词。
+      imageInsertPrompt: 图片描述插入到具体消息里的提示词。
+      gifStrategy:
+          $desc: GIF 图片帧提取策略
+          $inner:
+              - 只获取第一帧
+              - 获取前 N 帧
+              - 平均获取 N 帧
+
+      gifFrameCount: 提取的帧数（仅在选择"前 N 帧"或"平均 N 帧"策略时有效）
