+++ packages/service-multimodal/src/utils.ts
