+++ packages/service-multimodal/tsconfig.json
